// client/src/shared-module/model/koinly-tag.ts
export type KoinlyTag =
  | 'Airdrop'
  | 'Fork'
  | 'Mining'
  | 'Reward'
  | 'Salary'
  | 'Income'
  | 'Lending interest'
  | 'Cashback'
  | 'Fee refund'
  | 'Realized PNL'
  | 'Loan'
  | 'Margin loan'
  | 'Loan repayment'
  | 'Margin repayment'
  | 'Cost'
  | 'Tax'
  | 'Margin fee'
  | 'Loan fee'
  | 'Futures Fee'
  | 'Funding Fee'
  | 'Gift'
  | 'Lost'
  | 'Donation'
  | 'Swap'
  | 'Remove from Pool'
  | 'Add to Pool'
  | undefined;